pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "jfox-cli"
version = "0.1.4"
description = "JFox - Zettelkasten 知识管理 CLI 工具"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "User", email = "user@example.com"}
]
keywords = ["zettelkasten", "knowledge-management", "note-taking", "cli", "jfox"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "typer>=0.12.0",
    "rich>=13.0.0",
    "sentence-transformers>=3.0",
    "chromadb>=0.5.0",
    "networkx>=3.0",
    "watchdog>=3.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "rank-bm25>=0.2.2",
    "jinja2>=3.0",
]

[project.optional-dependencies]
dev = [
    # 基础测试
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-xdist>=3.0",       # 并行测试加速
    "pytest-timeout>=2.0",     # 超时控制
    
    # 代码质量
    "black>=23.0",
    "ruff>=0.1.0",
]

[project.scripts]
jfox = "jfox.cli:app"

[project.urls]
Homepage = "https://github.com/zhuxixi/jfox"
Repository = "https://github.com/zhuxixi/jfox"

[tool.hatch.build.targets.wheel]
packages = ["jfox"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501", "E402"]
